vector completed

+++ house.h
@@ -17,5 +17,5 @@ typedef struct house {
     boolean has_garbage_lift;
     uint32_t flats_count;
     uint32_t floors_count;
-    uint32_t avg_flat_square;
+    float avg_flat_square;
 } house;
